Refactoring Technique - Pull up method applied

+++ test/com/jwetherell/algorithms/mathematics/test/ExponentiationTest.java
@@ -19,7 +19,7 @@ public class ExponentiationTest {
         final List<Integer> expectedResultList = Arrays.asList(1, 134217728, 268435456, 362797056, 1073741824, 410338673, 7962624);
 
         for (int i = 0; i < expectedResultList.size(); i++)
-            assertEquals(expectedResultList.get(i), Exponentiation.recursiveExponentiation(baseList.get(i), exponentList.get(i)));
+            assertEquals(expectedResultList.get(i).intValue(), Exponentiation.recursiveExponentiation(baseList.get(i), exponentList.get(i)));
     }
 
     @Test
@@ -29,7 +29,7 @@ public class ExponentiationTest {
         final List<Integer> expectedResultList = Arrays.asList(1, 134217728, 268435456, 362797056, 1073741824, 410338673, 7962624);
 
         for (int i = 0; i < expectedResultList.size(); i++)
-            assertEquals(expectedResultList.get(i), Exponentiation.fastRecursiveExponentiation(baseList.get(i), exponentList.get(i)));
+            assertEquals(expectedResultList.get(i).intValue(), Exponentiation.fastRecursiveExponentiation(baseList.get(i), exponentList.get(i)));
     }
 
     @Test
@@ -40,6 +40,6 @@ public class ExponentiationTest {
         final List<Integer> expectedResultList = Arrays.asList(1, 2, 1, 0, 1, 8, 4, 2178);
 
         for (int i = 0; i < expectedResultList.size(); i++)
-            assertEquals(expectedResultList.get(i), Exponentiation.fastRecursiveExponentiationModulo(baseList.get(i), exponentList.get(i), divisorList.get(i)));
+            assertEquals(expectedResultList.get(i).intValue(), Exponentiation.fastRecursiveExponentiationModulo(baseList.get(i), exponentList.get(i), divisorList.get(i)));
     }
 }
